+++ api/src/pronunciation_api/word_metrics.py
@@ -3,7 +3,7 @@ import numpy as np
 
 
 # версия на чистом Python, не использующая внешних библиотек
-def edit_distance_python(a, b):
+def edit_distance_python(a: str, b: str):
     """
     Вычисляет расстояние Левенштейна между двумя строками `a` и `b`.
     Расстояние Левенштейна — это минимальное количество операций (вставка, удаление, замена),
